requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
eventlet==0.33.3
Flask==3.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.5
greenlet==3.3.2
h11==0.16.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6   
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
python-dotenv==1.0.0
python-engineio==4.13.1
python-socketio==5.10.0
simple-websocket==1.1.0
six==1.17.0
Werkzeug==3.0.1
wsproto==1.3.2
